+++ be/src/fs/fs_s3.cpp
@@ -22,8 +22,8 @@
 
 #include "common/config.h"
 #include "common/s3_uri.h"
+#include "fs/output_stream_adapter.h"
 #include "gutil/strings/util.h"
-#include "io/output_stream_adapter.h"
 #include "io/s3_input_stream.h"
 #include "io/s3_output_stream.h"
 #include "util/hdfs_util.h"
