+++ be/src/fs/output_stream_adapter.h
