+++ run-ut.sh
@@ -117,10 +117,14 @@ if [ ! -d ${CMAKE_BUILD_DIR} ]; then
     mkdir -p ${CMAKE_BUILD_DIR}
 fi
 
+# need config STAROS_DIR in custom.sh if USE_STAROS
+if [ "${USE_STAROS}" == "ON"  -a -z "${STAROS_DIR}" ]; then
+    echo "Please set STAROS_DIR when using staros"
+    exit 1
+fi
+
 cd ${CMAKE_BUILD_DIR}
 
-# need config STAROS_DIR in custom.sh if USE_STAROS
-export STAROS_DIR="/home/disk1/hanrui/dev/staros"
 if [ "${USE_STAROS}" == "ON"  ]; then
   ${CMAKE_CMD}  -G "${CMAKE_GENERATOR}" \
               -DSTARROCKS_THIRDPARTY=${STARROCKS_THIRDPARTY}\
@@ -223,8 +227,6 @@ cp -r ${STARROCKS_HOME}/be/test/util/test_data ${STARROCKS_TEST_BINARY_DIR}/util
 
 test_files=`find ${STARROCKS_TEST_BINARY_DIR} -type f -perm -111 -name "*test" | grep -v starrocks_test | grep -v bench_test`
 
-echo "    ${STARROCKS_TEST_BINARY_DIR}/starrocks_test --gtest_filter=${TEST_FILTER}"
-/home/disk1/sr-deps/toolchain/installed/gdb-10.2/bin/gdb ${STARROCKS_TEST_BINARY_DIR}/starrocks_test
 # run cases in starrocks_test in parallel if has gtest-parallel script.
 # reference: https://github.com/google/gtest-parallel
 if [ -x ${GTEST_PARALLEL} ]; then
